Add files via upload
("GEORGE JUNIOR MAGAZINE Official Questionnaire Website")  
_A Digital Tribute to JFK Sr. & JFK Jr._  

This repository hosts the official questionnaire website dedicated to the legacy of John F. Kennedy Sr. and John F. Kennedy Jr., celebrating their contributions and vision. The project aims to engage the community through thought-provoking questions, discussions, and historical reflections. Explore interactive features and join in preserving the values and ideals of two iconic figures in American history. Contributions are welcome as we build a platform for dialogue and remembrance.

🔗 **Follow us on:**
- [YouTube](https://www.youtube.com/@georgejrmagazine)
- [Quora](https://www.quora.com/profile/George-Jr-Magazine?ch=17&oid=2048399969&share=6f38f659&srid=hcOUcW&target_type=user)